+++ CHANGELOG.md
@@ -3,6 +3,7 @@ In Development
 - Reduced the sizes of a number of streams & futures
 - Added doc comments to much of the code
 - Return 502 status when a backend returns an invalid response
+- Require `--api-url` (and other URLs retrieved from APIs) to be HTTP(S)
 
 v0.4.0 (2024-07-09)
 -------------------